Feature/various updates

+++ thumbservice.py
@@ -1,43 +1,174 @@
 #!/usr/bin/env python
-from flask import Flask, request, abort, jsonify, redirect
-from flask_cors import CORS
-import requests
 import os
+import uuid
+
 import boto3
+import logging
+import requests
+from flask_cors import CORS
+from flask.logging import default_handler
+from flask import Flask, request, jsonify, redirect
 from fits2image.conversions import fits_to_jpg
 from fits_align.ident import make_transforms
 from fits_align.align import affineremap
+
 app = Flask(__name__)
 CORS(app)
 
-ARCHIVE_API = 'https://archive-api.lco.global/'
-TMP_DIR = os.getenv('TMP_DIR', '/tmp/')
-BUCKET = os.getenv('AWS_S3_BUCKET', 'lcogtthumbnails')
+
+class RequestFormatter(logging.Formatter):
+    def format(self, record):
+        record.url = request.url
+        return super().format(record)
+
+formatter = RequestFormatter('[%(asctime)s] %(levelname)s in %(module)s for %(url)s: %(message)s')
+default_handler.setFormatter(formatter)
+
+
+class Settings:
+    def __init__(self, settings=None):
+        self._settings = settings or {}
+
+        self.ARCHIVE_API = self.set_value('ARCHIVE_API', 'https://archive-api.lco.global/', True)
+        self.TMP_DIR = self.set_value('TMP_DIR', '/tmp/', True)
+        self.BUCKET = self.set_value('AWS_S3_BUCKET', 'lcogtthumbnails')
+        self.AWS_ACCESS_KEY_ID = self.set_value('AWS_ACCESS_KEY_ID', 'changeme')
+        self.AWS_SECRET_ACCESS_KEY = self.set_value('AWS_SECRET_ACCESS_KEY', 'changeme')
+        # Using `None` for `STORAGE_URL` will connect to AWS
+        self.STORAGE_URL = self.set_value('STORAGE_URL', None)
+
+    def set_value(self, env_var, default, must_end_with_slash=False):
+        if env_var in self._settings:
+            value = self._settings[env_var]
+        else:
+            value = os.getenv(env_var, default)
+        return self.end_with_slash(value) if must_end_with_slash else value
+
+    @staticmethod
+    def end_with_slash(path):
+        return os.path.join(path, '')
+
+settings = Settings()
+
+
+class ThumbnailAppException(Exception):
+    status_code = 500
+
+    def __init__(self, message, status_code=None, payload=None):
+        Exception.__init__(self)
+        self.message = message
+        if status_code is not None:
+            self.status_code = status_code
+        self.payload = payload
+
+    def to_dict(self):
+        result = dict(self.payload or ())
+        result['message'] = self.message
+        return result
+
+
+@app.errorhandler(ThumbnailAppException)
+def handle_thumbnail_app_exception(error):
+    response = jsonify(error.to_dict())
+    response.status_code = error.status_code
+    return response
+
+
+@app.errorhandler(Exception)
+def handle_broad_exceptions(error):
+    app.logger.error(error, exc_info=True)
+    response = jsonify({'message': 'Internal server error'})
+    response.status_code = 500
+    return response
+
+
+def get_response(url, params=None, headers=None):
+    response = None
+    try:
+        response = requests.get(url, headers=headers, params=params)
+        response.raise_for_status()
+    except requests.RequestException:
+        status_code = getattr(response, 'status_code', None)
+        payload = {}
+        message = 'Got error response'
+        if status_code is None or 500 <= status_code < 600:
+            status_code = 502
+        elif status_code == 404:
+            message = 'Not found'
+        else:
+            try:
+                payload['response'] = response.json()
+            except:
+                pass
+        raise ThumbnailAppException(message, status_code=status_code, payload=payload)
+    return response
+
+
+def get_unique_id():
+    return uuid.uuid4().hex
+
+
+def can_generate_thumbnail_on(frame, request):
+    frame_has_required_validation_keys = all([key in frame.keys() for key in ['OBSTYPE', 'REQNUM', 'filename']])
+    if not frame_has_required_validation_keys:
+        return {'result': False, 'reason': 'Cannot generate thumbnail for given frame'}
+
+    valid_obstypes = [
+        'ARC', 'BIAS', 'BPM', 'DARK', 'DOUBLE', 'EXPERIMENTAL', 'EXPOSE', 'GUIDE', 'LAMPFLAT', 'SKYFLAT',
+        'SPECTRUM', 'STANDARD', 'TARGET', 'TRAILED'
+    ]
+    valid_obstypes_for_color_thumbs = ['EXPOSE', 'STANDARD']
+    obstype = frame.get('OBSTYPE').upper()
+    reqnum = frame.get('REQNUM')
+    is_color_request = request.args.get('color', 'false') == 'true'
+    is_fits_file = any([frame.get('filename').endswith(ext) for ext in ['.fits', '.fits.fz']])
+
+    if obstype not in valid_obstypes:
+        return {'result': False, 'reason': f'Cannot generate thumbnail for OBSTYPE={obstype}'}
+
+    if is_color_request and not reqnum:
+        return {'result': False, 'reason': 'Cannot generate color thumbnail for a frame that does not have a request'}
+
+    if is_color_request and obstype not in valid_obstypes_for_color_thumbs:
+        return {'result': False, 'reason': f'Cannot generate color thumbnail for OBSTYPE={obstype}'}
+
+    if not is_fits_file:
+        return {'result': False, 'reason': 'Cannot generate thumbnail for non FITS-type frame'}
+
+    return {'result': True, 'reason': ''}
 
 
 def save_temp_file(frame):
-    path = TMP_DIR + frame['filename']
+    path = f'{settings.TMP_DIR}{get_unique_id()}-{frame["filename"]}'
     with open(path, 'wb') as f:
-        f.write(requests.get(frame['url']).content)
+        f.write(get_response(frame['url']).content)
     return path
 
 
 def key_for_jpeg(frame_id, **params):
-    return '{0}.{1}.jpg'.format(frame_id, hash(frozenset(params.items())))
+    return f'{frame_id}.{hash(frozenset(params.items()))}.jpg'
 
 
 def convert_to_jpg(paths, key, **params):
-    jpg_path = os.path.dirname(paths[0]) + '/' + key
+    jpg_path = f'{os.path.dirname(paths[0])}/{get_unique_id()}-{key}'
     fits_to_jpg(paths, jpg_path, **params)
     return jpg_path
 
 
-def upload_to_s3(jpg_path):
-    key = os.path.basename(jpg_path)
-    client = boto3.client('s3')
+def get_s3_client():
+    return boto3.client(
+        's3',
+        aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
+        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
+        endpoint_url=settings.STORAGE_URL
+    )
+
+
+def upload_to_s3(key, jpg_path):
+    client = get_s3_client()
     with open(jpg_path, 'rb') as f:
         client.put_object(
-            Bucket=BUCKET,
+            Bucket=settings.BUCKET,
             Body=f,
             Key=key,
             ContentType='image/jpeg'
@@ -45,38 +176,29 @@ def upload_to_s3(jpg_path):
 
 
 def generate_url(key):
-    client = boto3.client('s3')
+    client = get_s3_client()
     return client.generate_presigned_url(
         'get_object',
         ExpiresIn=3600 * 8,
-        Params={'Bucket': BUCKET, 'Key': key}
+        Params={'Bucket': settings.BUCKET, 'Key': key}
     )
 
 
 def key_exists(key):
-    client = boto3.client('s3')
+    client = get_s3_client()
     try:
-        client.head_object(Bucket=BUCKET, Key=key)
+        client.head_object(Bucket=settings.BUCKET, Key=key)
         return True
     except:
         return False
 
 
-def frames_for_requestnum(reqnum, request):
+def frames_for_requestnum(reqnum, request, rlevel):
     headers = {
         'Authorization': request.headers.get('Authorization')
     }
-    frames = requests.get(
-        '{0}frames/?REQNUM={1}'.format(ARCHIVE_API, reqnum),
-        headers=headers
-    ).json()['results']
-    if any(f for f in frames if f['RLEVEL'] == 91):
-        rlevel = 91
-    elif any(f for f in frames if f['RLEVEL'] == 11):
-        rlevel = 11
-    else:
-        rlevel = 0
-    return [f for f in frames if f['RLEVEL'] == rlevel]
+    params = {'REQNUM': reqnum, 'RLEVEL': rlevel}
+    return get_response(f'{settings.ARCHIVE_API}frames/', params=params, headers=headers).json()['results']
 
 
 def rvb_frames(frames):
@@ -85,7 +207,6 @@ def rvb_frames(frames):
         'visual': ['V'],
         'blue': ['B'],
     }
-
     selected_frames = []
     for color in ['red', 'visual', 'blue']:
         try:
@@ -93,22 +214,47 @@ def rvb_frames(frames):
                 next(f for f in frames if f['FILTER'] in FILTERS[color])
             )
         except StopIteration:
-            abort(404)
+            raise ThumbnailAppException('RVB frames not found', status_code=404)
     return selected_frames
 
-def reproject_files(ref_image, images_to_align):
-    identifications = make_transforms(ref_image, images_to_align[1:3])
 
+def reproject_files(ref_image, images_to_align):
     aligned_images = []
-    for id in identifications:
-        if id.ok:
-            aligned_img = affineremap(id.ukn.filepath, id.trans, outdir=TMP_DIR)
-            aligned_images.append(aligned_img)
+
+    try:
+        identifications = make_transforms(ref_image, images_to_align[1:3])
+        for id in identifications:
+            if id.ok:
+                aligned_img = affineremap(id.ukn.filepath, id.trans, outdir=settings.TMP_DIR)
+                aligned_images.append(aligned_img)
+    except Exception:
+        app.logger.error('Error aligning images, falling back to original image list', exc_info=True)
+        # Clean up any images that were created
+        for image in aligned_images:
+            if os.path.exists(image):
+                os.remove(image)
 
     img_list = [ref_image]+aligned_images
     if len(img_list) != 3:
         return images_to_align
     return img_list
+
+
+class Paths:
+    """Retain all paths set"""
+    def __init__(self):
+        self._all_paths = set()
+        self.paths = []
+
+    def set(self, paths):
+        for path in paths:
+            self._all_paths.add(path)
+        self.paths = paths
+
+    @property
+    def all_paths(self):
+        return list(self._all_paths)
+
 
 def generate_thumbnail(frame, request):
     params = {
@@ -125,26 +271,32 @@ def generate_thumbnail(frame, request):
         return generate_url(key)
     # Cfitsio is a bit crappy and can only read data off disk
     jpg_path = None
-    paths = []
+    paths = Paths()
     try:
         if not params['color']:
-            paths = [save_temp_file(frame)]
+            paths.set([save_temp_file(frame)])
         else:
-            reqnum_frames = frames_for_requestnum(frame['REQNUM'], request)
-            paths = [save_temp_file(frame) for frame in rvb_frames(reqnum_frames)]
-            paths = reproject_files(paths[0], paths)
-        jpg_path = convert_to_jpg(paths, key, **params)
-        upload_to_s3(jpg_path)
+            # Color thumbnails can only be generated on rlevel 91 images
+            reqnum_frames = frames_for_requestnum(frame['REQNUM'], request, rlevel=91)
+            paths.set([save_temp_file(frame) for frame in rvb_frames(reqnum_frames)])
+            paths.set(reproject_files(paths.paths[0], paths.paths))
+        jpg_path = convert_to_jpg(paths.paths, key, **params)
+        upload_to_s3(key, jpg_path)
     finally:
         # Cleanup actions
-        if jpg_path:
+        if jpg_path and os.path.exists(jpg_path):
             os.remove(jpg_path)
-        for path in paths:
-            os.remove(path)
+        for path in paths.all_paths:
+            if os.path.exists(path):
+                os.remove(path)
     return generate_url(key)
 
 
 def handle_response(frame, request):
+    can_generate_thumbnail_on_frame = can_generate_thumbnail_on(frame, request)
+    if not can_generate_thumbnail_on_frame['result']:
+        raise ThumbnailAppException(can_generate_thumbnail_on_frame['reason'], status_code=400)
+
     url = generate_thumbnail(frame, request)
     if request.args.get('image'):
         return redirect(url)
@@ -157,12 +309,12 @@ def bn_thumbnail(frame_basename):
     headers = {
         'Authorization': request.headers.get('Authorization')
     }
-    frames = requests.get(
-        '{0}frames/?basename={1}'.format(ARCHIVE_API, frame_basename),
-        headers=headers
-    ).json()
-    if not 0 < frames['count'] < 2:
-        abort(404)
+    params = {'basename': frame_basename}
+    frames = get_response(f'{settings.ARCHIVE_API}frames/', params=params, headers=headers).json()
+
+    if not frames['count'] == 1:
+        raise ThumbnailAppException('Not found', status_code=404)
+
     return handle_response(frames['results'][0], request)
 
 
@@ -171,12 +323,8 @@ def thumbnail(frame_id):
     headers = {
         'Authorization': request.headers.get('Authorization')
     }
-    frame = requests.get(
-        '{0}frames/{1}/'.format(ARCHIVE_API, frame_id),
-        headers=headers
-    ).json()
-    if frame.get('detail') == 'Not found.':
-        abort(404)
+    frame = get_response(f'{settings.ARCHIVE_API}frames/{frame_id}/', headers=headers).json()
+
     return handle_response(frame, request)
